Add functionality to generate random posts for a user
This commit introduces a new endpoint in the PostsController (`POST /posts/random`) which allows authenticated users to generate a predefined number of random posts. The endpoint uses an AccessTokenGuard to ensure that the user is authenticated before proceeding. The PostsService has been extended with a new `generatePosts` method that creates 100 posts with predetermined titles and contents. A temporary TODO comment has been added to remind us to remove or adjust this test functionality later.

A new import statement for `UsersModel` from the `src/users/entities/users.entity` module has been added to the PostsController to support user data retrieval within the new endpoint.